Use a faster algorithm to find the subgraph
The current algorithm is O(START*END*(V + E)), which can be quite
expensive. Instead, we can't use only two BFSs: one from all the START
nodes and another from the END nodes. Then, the result will come from
the intersection of the two sets of reachable nodes. This drops the
complexity down to O(V + E).

For the search, we don't even need a BFS as the order of node visiting
is not important, so we use a basic search with an unordered
neighbourhood.

Suggested-by: Pedro Sousa <pedroteosousa@gmail.com>
Helped-by: Mateus Barbosa <mcmfbarbosa@gmail.com>
Signed-off-by: Lucas Oshiro <lucasseikioshiro@gmai.com>
Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>