Recomend similar function names for not found ones
Sometimes, when using the plugin with a Makefile, specifying a function
as "base_filename:func" may not work because the Makefile invokes gcc
with a filename such as "./base_filename" or "<dir>/base_filename". This
makes the plugin fail to find the function, with an error message. But
the user may have a hard time getting to know how they should specify
the function. So let's compare the given function specification with the
ones we know and recomend the most similar ones to the user.

Closes #2

Signed-off-by: Matheus Tavares <matheus.bernardino@usp.br>